Add Picker mock
