--- web/src/mocks/Picker.js
@@ -1,3 +1,0 @@
-const Picker = () => null;
-Picker.Item = () => null;
-export default Picker;
